arrow_scraper/requirements.txt
crawl4ai>=0.7.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
asyncio>=3.4.3
aiohttp>=3.8.0
Flask==2.3.3
Flask-CORS>=4.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
click>=8.1.7
itsdangerous>=2.1.2
MarkupSafe>=2.1.3